chore: shut linter up

+++ internal/dbutil/list_query.go
@@ -61,7 +61,7 @@ func (q ListQuery) Build() (string, []any, error) {
 	}
 	args := slices.Clone(q.Args)
 	limitIndex := len(args) + 1
-	args = append(args, int32(params.PageSize), int32(params.PageIndex*params.PageSize))
+	args = append(args, params.PageSize, params.PageIndex*params.PageSize)
 
 	parts := q.baseParts()
 	if orderSQL != "" {
